+++ src/app/app.cc
@@ -5,6 +5,7 @@
 
 #include "app/window.h"
 #include "app/app_settings.h"
+#include "engine/engine.h"
 
 
 namespace petri {
@@ -19,11 +20,15 @@ int App::Run() {
 	if (!window.Create(settings.windowSettings)) {
 		return -1;
 	}
+
+	engine::Engine engine;
+
 	while (!window.ShouldClose())
 	{
+		window.PollEvents();
+		engine.Tick();
 		window.ClearColor();
 		window.SwapBuffers();
-		window.PollEvents();
 	}
 	return 0;
 }